Add files via upload
House Grade Prediction
Find The Best House : The 1234ABC Realtors limited is in great losses due to the autmoation in the industry and the new innovations and players in the real estate market . So they have hired a Data Science Team to give them a business solution so as to help regain their place in the market. In order to do that , the group thought of a compiled way of grading the houses inorder from a to e with respect to different aspects in mind. This would help them to rate the houses as well as different areas as well. So gear up and help the realtors get back their place in the market.

Objective of the problem: The objective of the problem is to predict values “Grade” attribute from the given features of the Test data.
Evaluation: The metric of evaluation for this prediction problem is Accuracy.